* fhandler_serial.cc (fhandler_serial::raw_read): Use correct type for minchars. (fhandler_serial::ioctl): Set errno if the ClearCommError fails. (fhandler_serial::tcsetattr): Use correct value for vmin_. (fhandler_serial::tcgetattr): Ditto.

+++ winsup/cygwin/fhandler.h
@@ -626,7 +626,7 @@ class fhandler_cygdrive: public fhandler_disk_file
 class fhandler_serial: public fhandler_base
 {
  private:
-  unsigned int vmin_;			/* from termios */
+  size_t vmin_;				/* from termios */
   unsigned int vtime_;			/* from termios */
   pid_t pgrp_;
   int rts;				/* for Windows 9x purposes only */
